pyproject.toml
[project]
name = "pptx-slide-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "icrawler>=0.6.10",
    "markitdown[docx]>=0.1.5",
    "pillow>=12.1.1",
    "pywin32>=311; sys_platform == 'win32'",
    "python-pptx>=1.0.2",
    "svglib>=1.6.0",
    "reportlab>=4.4.10",
    "kiwisolver>=1.5.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "notebooklm-py>=0.3.4",
]

[tool.ruff]
target-version = "py313"
exclude = ["temp", "ref"]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
